Create release workflow
When a change is pushed to the main branch of the original repository
that includes Terraform or Python changes, a new GitHub Release and git
tag are created. Their number and description are extracted from the
CHANGELOG.

The workflow may be triggered manually to create historical releases.
However, in this case, the commit must exist on the main branch, which
must also describe the version in the CHANGELOG.